added useState to handle forms input
